requirements.txt
langchain-openai==0.3.24
langchain==0.3.26
langchain-community==0.3.26
langchain-text-splitters==0.3.8
playwright==1.52.0
python-dotenv==1.1.0
streamlit==1.38.0
pydantic>=2.0.0
requests==2.32.4
